this checkin has the implementation for iperf. It has the following main classes:
Listener Class
IperfUtils
IperfStats
IperfErrors
Iperf

ParseIperfStats is the function that parses the output of iperf and gets the bandwidth as well as the units.
The iperf class assumes that a server has been started with the iperf -s command on a seperate computer.

The Main Activity class has been modified to just invoke the iperf functionality. It needs to be modified to include the cellular and network manager code.