[WIP] Migrate the NDK dependencies to use cmake
CMake is recommended for new projects currently, and has more
sophisticated support for external build systems than
ndk-build. Ultimately it will be good to build iperf3 with their own
included build system (based on autotools) while still integrating the
iperf3 code into the codebase for tracing and debugging.

This commit restructures the repo layout to match the recommended
android cmake layout, and re-creates the existing ndk-build toolchain
in CMake. This lays the foundation for  better integrating modern
iperf in a forthcoming commit.