Refactor the iperf3 integration to an overall wrapper
The wrapper should be responsible for rectifying the lifecycle of the
underlying iperf test with the asynchronous requests and reports
to/from the java side of the interface. Importantly the wrapper should
not touch the internals of iperf, relying on the iperf API to make it
easier to upgrade/change the iperf library in the future if needed.