Determine the caching directory programatically in java
The android context API provides a direct means to identify the cache
directory for the application. The previous approach inferred the
correct directory based off the app name, but this does not work in
the android emulator, at least at API level 30. This commit changes
the logic to pass the cache directory template from the
Iperf3Client.java class into the native executor, keeping the
business logic and policy in java as much as possible.