Return synchronous run result via JNI
This commit modifys the run function to directly return a status code,
giving a clear opportunity for the run function to indicate it has
completed and the work task has either succeeded or failed. This
simplifies the call flow and eliminates the need to cancel the test
within the error handler.