Ensure only one iperf runs at a time
Currently the iperf code relies on static resources within the
process (specifically a pointer to the current underlying test
structure). Since the "stop" signal can come from a different thread
context and will need to be mapped to a single running iperf, more
work would be required to support multiple concurrent operations. This
commit adds mutexes to the wrapper class to ensure that only one iperf
is running at a time, and that the stop signal is applied to a valid
test structure that has not been concurrently freed.

TECHNICALLY THERE IS STILL A RACE CONDITION IN THIS CODE, since both
the running iperf test and the stop signal can concurrently modify the
"done" flag. Ideally mutex coordination could be added to iperf itself
to synchronize access to the flag, or the flag could be moved to a
synchronized atomic or semaphore.