Synchronize iperf worker shutdown
It appears that the WorkManager will call onStopped, and at the point
onStopped returns, will cancel the task. Task cancellation appears to
be stopping the executing iperf, but causes the thread to stop
uncleanly and leak lock resources. This commit adds an explicit signal
from the doWork context to the onStopped context to allow onStopped to
wait until the worker has cleaned up the main thread. It currently has
a timeout of 10s to abort the thread uncleanly if the iperf library
hangs.