Set a global iperf wrapper instance to allow out-of-band stop
To avoid passing ownership across the NDK boundary, this commit sets
up a global pointer on the cpp side to a single instance of the
IperfStateWrapper that should be running at a time. This approach will
break if multiple instances of Iperf exec are ever run at the same
time, but allows for cleanly managing the lifetimes of the Iperf
wrapper instance from multiple thread contexts. This supports the
interrupt strategy used by the java iperf worker, which will call
stopIperfTest from a different thread context than the running main
task (which will be synchronously blocking the main thread).