Handle the edge case when a test is stopped while still initializing
Within libiperf iperf_client_api.c:run_iperf_client will reset the
"done" flag to 0 while handshaking with the server. In our
asynchronous case it is possible for a user to cancel the test before
this occurs. Previously this would result in the "done" flag begin set
to 1, then reset back to 0 before the iperf client actually begins
running!

This commit adds a global flag in the context of the wrapper to
indicate that the global test has been requested to stop. Only at the
point an IperfTestWrapper is destructed is this flag set to
false. This allows it to be checked at periodic intervals (in the
periodic callback functions) to fix the state of the "done" flag in
the actual iperf_test struct.

This still feels pretty hacky, and it would be better if the iperf
client api had a cleaner way to manage the client lifetime.